Refactor transport parameters code into separate module file.
